requirements.txt
openai-whisper
yt-dlp
langchain
faiss-cpu
openai
streamlit
tiktoken
python-dotenv
langchain-community